Добавляет модули data.js и util.js, подключает main.js, как точку входа
